git-svn-id: http://svn.castleproject.org:8080/svn/castle/trunk@373 73e77b4c-caa6-f847-a29a-24ab75ae54b6

+++ Facilities/Prevalence/Castle.Facilities.Prevalence/PrevalenceEngineComponentActivator.cs
@@ -46,7 +46,7 @@ namespace Castle.Facilities.Prevalence
 				Model.ExtendedProperties[PrevalenceFacility.AutoMigrationPropertyKey];
 			bool resetStorage = (bool) 
 				Model.ExtendedProperties[PrevalenceFacility.ResetStoragePropertyKey];
-			int snapshotPeriod = (Int32)Model.ExtendedProperties[PrevalenceFacility.SnapshotPeriodPropertyKey];
+			float snapshotPeriod = (float)Model.ExtendedProperties[PrevalenceFacility.SnapshotPeriodPropertyKey];
 			
 			if (resetStorage)
 			{
@@ -67,7 +67,7 @@ namespace Castle.Facilities.Prevalence
 			return engine;
 		}
 
-		private void CreateSnapshotTaker(PrevalenceEngine engine, int snapshotPeriod)
+		private void CreateSnapshotTaker(PrevalenceEngine engine, float snapshotPeriod)
 		{
 			TimeSpan period = TimeSpan.FromHours(snapshotPeriod);
 			ICleanUpPolicy policy = (ICleanUpPolicy) Kernel[PrevalenceFacility.CleanupPolicyComponentPropertyKey];
